move themisFlags to RunThemis
